+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Configure pixi in `pyproject.toml`
 - Handle GLIBC version incompatibility in `start_simulation.sh`
+- Spine: Throw an exception when a servo reply has invalid torque
 - Update downloaded simulation spine in cache if outdated
 - actuation: Added link inertia randomization to the Bullet interface (thanks to @Tordjx)
 - actuation: Check maximum torques before sending commands